requirements.txt
python-gitlab==4.4.0
python-dotenv==1.0.1

reportlab==4.1.0

pytest==8.0.0
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-reporter-html1==0.8.3

mkdocs==1.5.3
mkdocs-material==9.5.8

numpy==1.26.4
pandas==2.2.0
pyarrow==15.0.0
matplotlib==3.8.2
seaborn==0.13.1

dash==2.14.2
dash[testing]==2.14.2

typer[all]==0.9.0
